ay/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
django-ckeditor==6.7.0
django-extensions==3.2.1
django-js-asset==2.1.0
django-modeltranslation==0.18.10
django-parler==2.3
django-rosetta==0.9.9
idna==3.4
mysqlclient==2.2.0
Pillow==9.4.0
polib==1.2.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.6.3
urllib3==2.0.3
